Phase 2: move gates/parser/compiler/registry/topology/qft/trotter into dense_evolution.circuits/
- git mv all 7 modules into dense_evolution/circuits/ (largest subpackage
  split so far, heavy internal cross-dependencies)
- new circuits/__init__.py re-exports the public API
- backward-compat shims at all 7 old top-level paths
- fixed relative-import fallout across the codebase: dense_evolution/autodiff.py,
  mps.py, simulator.py (all still flat root, not moving until Phase 2's
  backends/solvers steps) now import gates/parser/compiler/registry via
  .circuits.<mod>; dense_evolution/interop/qiskit_pennylane.py's ..parser ->
  ..circuits.parser; dense_evolution/physics/states.py's ..topology ->
  ..circuits.topology
- gates.py's own internal `.registry` import needed no change: both modules
  move into circuits/ together, so the relative import still resolves correctly
- updated dense_evolution/__init__.py, the 7 docs/api/*.md mkdocstrings
  directives, and pyproject.toml packages list

Fixed one real regression found during verification: tests/test_simulator.py
imported private JIT helpers (_compile_and_run_circuit_jit(_donated)) from
the dense_evolution.compiler shim, which (correctly) only re-exports the
public API -- pointed that import at dense_evolution.circuits.compiler
directly instead, same pattern already used for the interop/ PR's private
test hooks.

Also identified, but deliberately left unfixed as out of scope for a pure
file-move PR: a pre-existing (not introduced by this move) Python
self-shadowing landmine on dense_evolution.qft. The function and its
defining module share the exact name "qft", so any code path that does
`from dense_evolution.qft import qft` or `import dense_evolution.qft`
after dense_evolution was already imported silently overwrites the
dense_evolution.qft package attribute from the function back to the
submodule -- breaking any later `de.qft(...)` call (e.g.
tools/dashboard_core/qasm_library.py's module-level QASM_LIBRARY dict
construction) in the same process. Confirmed this identical risk already
existed before Phase 2 (same self-named function/module collision at the
old flat dense_evolution/qft.py), and confirmed it does not manifest in
the real test suite because pytest's default alphabetical collection runs
test_mps.py before test_qft.py. Worth a real fix later (e.g. renaming the
function or restructuring the re-export), tracked as a follow-up.

Full suite: 796 passed, 17 skipped, 4 failed on first run -- all 4 were
MemoryPressureError (known resource-contention pattern), confirmed via
isolated re-run: all 4 pass individually.

Part of #76.
